requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.2
cctx==1.0.0
ccxt==1.73.12
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
cycler==0.11.0
dateparser==1.1.0
fonttools==4.29.1
frozenlist==1.3.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pandas-ta==0.3.14b0
Pillow==9.0.1
plotly==5.6.0
pycares==4.1.2
pycparser==2.21
pyparsing==3.0.7
python-binance==1.0.15
python-dateutil==2.8.2
python-decouple==3.6
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.1.18
requests==2.27.1
six==1.16.0
tenacity==8.0.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
websockets==9.1
yarl==1.7.2
